+++ src/main/java/io/github/mirancz/gtfsparser/util/StopInfo.java
@@ -1,9 +1,8 @@
 package io.github.mirancz.gtfsparser.util;
 
-import java.io.DataOutputStream;
 import java.io.IOException;
 
-public record StopInfo(int stopId, int postId) {
+public record StopInfo(short stopId, short postId) {
 
     public void write(CheckedOutputStream os) throws IOException {
         os.writeShort(stopId);
